// ClientApp/src/store/toolbox/toolbox.types.ts
export enum TOOLBOX_ACTION_TYPES {
    SET_IS_ARTIFICIALINTELLIGENCE_OPEN = 'toolbox/SET_IS_ARTIFICIALINTELLIGENCE_OPEN',
    SET_IS_VITALS_OPEN = 'toolbox/SET_IS_VITALS_OPEN',
    SET_IS_POST_OPEN = 'toolbox/SET_IS_POST_OPEN',
    SET_IS_SOLARSYSTEM_OPEN = 'toolbox/SET_IS_SOLARSYSTEM_OPEN',
    SET_IS_PLANETS_OPEN = 'toolbox/SET_IS_PLANETS_OPEN',
    SET_IS_MOONS_OPEN = 'toolbox/SET_IS_MOONS_OPEN',
    SET_IS_MARAUDERS_OPEN = 'toolbox/SET_IS_MARAUDERS_OPEN',
    SET_IS_MESSAGES_OPEN = 'toolbox/SET_IS_MESSAGES_OPEN'
}